Reverse "forced" mu-plugins & "classic" mu-plugins. Ask before installing each classic mu-plugin.

+++ bin/mu.sh
@@ -8,18 +8,20 @@ echo '### MU-Plugins installation';
 
 mkdir "${MAINDIR}${WP_MUPLUGINS_DIR}";
 
-# Classic MU Plugins
-for i in $WPU_MUPLUGINS
+# Forced MU Plugins
+for i in $WPU_FORCED_MUPLUGINS
 do
     echo "## Install ${i}";
     cp "${MAINDIR}WPUtilities/${WP_MUPLUGINS_DIR}${i}.php" "${MAINDIR}${WP_MUPLUGINS_DIR}${i}.php";
 done;
 
-# Forced MU Plugins
-for i in $WPU_FORCED_MUPLUGINS
+# Classic MU Plugins
+for i in $WPU_MUPLUGINS
 do
-    echo "## Install ${i}";
-    cp "${MAINDIR}WPUtilities/${WP_PLUGINS_DIR}${i}.php" "${MAINDIR}${WP_MUPLUGINS_DIR}${i}.php";
+    read -p "## Install ${i} ? (y/N) " install_muplugin;
+    if [[ $install_muplugin == 'y' ]];then
+        cp "${MAINDIR}WPUtilities/${WP_PLUGINS_DIR}${i}.php" "${MAINDIR}${WP_MUPLUGINS_DIR}${i}.php";
+    fi;
 done;
 
 # Commit Add mu-plugins
